+++ compiler/rustc_codegen_llvm/src/gotoc/mir_to_goto/mod.rs
@@ -8,9 +8,8 @@ mod backend;
 mod codegen;
 mod context;
 mod debug;
-mod hooks;
 mod monomorphize;
-mod stubs;
+mod overrides;
 mod utils;
 
 pub use backend::GotocCodegenBackend;
