+++ compiler/rustc_codegen_llvm/src/gotoc/mir_to_goto/overrides/stubs/hash_map_stub.rs
